+++ src/CommonMarkConverter.php
@@ -36,4 +36,11 @@ final class CommonMarkConverter extends MarkdownConverter
 
         parent::__construct($environment);
     }
+
+    public function getEnvironment(): Environment
+    {
+        \assert($this->environment instanceof Environment);
+
+        return $this->environment;
+    }
 }