+++ src/GithubFlavoredMarkdownConverter.php
@@ -35,4 +35,11 @@ final class GithubFlavoredMarkdownConverter extends MarkdownConverter
 
         parent::__construct($environment);
     }
+
+    public function getEnvironment(): Environment
+    {
+        \assert($this->environment instanceof Environment);
+
+        return $this->environment;
+    }
 }